Make the brcm ldisc driver compatible with libbt (use it for BT)
Kernel changes that allow to use the broadcom ldisc driver for combined
BT and FMRadio chips for BT using libbt. The major modification is the
interception of the hci configuration commands emitted by libbt and
their simulated answer to the upper layers. The configuration is done
already by the driver, while a reconfiguration would break the
functionality of the ldisc driver.
This modification requires that libbt is configured to communicate via
the ldisc's device node ("/dev/brcm_bt_drv"), while uim is wired to the
"real" uart port.

This change includes the following additional and required changes:
* "ifdef" the bluesleep start/stop hooks in the rfkill driver since it
is handled in the ldisc driver.
* Only reset the outgoing data timer (bluesleep) if the brcm driver is
not used. Also handled in the ldisc driver.
* msm_serial_hs: reset the client counter to 0 only if the ldisc driver
is not used since it is invoked twice when starting the ldisc, resulting
in a mismatching counter.

Required userspace modifications:
* "#define HCILP_INCLUDED false" in bdroid_buildcfg.h
* "UartPort = /dev/brcm_bt_dev" in bt_vendor.conf
* "UimUartPort = /dev/ttyHS0" in bt_vendor.conf (for msm_serial_hs)

Change-Id: Ib1a551b47280d3881f71d12b0dcfd274ff7efc4b
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>
(cherry picked from commit c5c7550a584a56225b834525e2f0d014a1cbdaea)